+++ src/components/IncidentDetails/select-components.js
@@ -1,12 +1,6 @@
 import { components } from "react-select";
-import {
-  HiOutlineChevronDoubleUp,
-  HiOutlineChevronDown,
-  HiOutlineChevronUp
-} from "react-icons/hi";
 import React from "react";
 import { Avatar } from "../Avatar";
-import { IncidentPriority } from "../../constants/incident-priority";
 
 export const IncidentPrioritySingleValue = ({ children, ...props }) => (
   <components.SingleValue {...props}>
@@ -57,26 +51,3 @@ export const IncidentRespondentsOption = ({ ...props }) => (
     </div>
   </components.Option>
 );
-export const priorities = [
-  {
-    label: "Low",
-    value: IncidentPriority.Low,
-    icon: function IconToProps(props) {
-      return <HiOutlineChevronDown color="green" {...props} />;
-    }
-  },
-  {
-    label: "Medium",
-    value: IncidentPriority.Medium,
-    icon: function IconToProps(props) {
-      return <HiOutlineChevronUp color="red" {...props} />;
-    }
-  },
-  {
-    label: "High",
-    value: IncidentPriority.High,
-    icon: function IconToProps(props) {
-      return <HiOutlineChevronDoubleUp color="red" {...props} />;
-    }
-  }
-];
